Added 120Hz refresh rate option for new ProMotion MacBook Pros. Fixes #1.
I don't actually have one of the shiny new MacBook Pros, but I do have a 144Hz external monitor, which I set to 120Hz, and it worked fine on that.
Also added 90 and 144 options.

Note: it is recommended that you add a 120Hz custom resolution, in the Nvidia control panel, and set your desktop resolution to that resolution, in order to get an optimal streaming experience.